+++ .gitignore
@@ -1,4 +1,5 @@
 .idea
+dist/
 pypi_cache
 __pycache__
 .pytest_cache
