+++ docs/source/conf.py
@@ -19,7 +19,8 @@ release = "0.0.1"
 import os
 import sys
 
-sys.path.insert(0, os.path.abspath("../.."))
+sys.path.insert(0, os.path.abspath("../../req2flatpak"))
+sys.path.insert(0, os.path.abspath("../../tests"))
 
 
 # =============================================================================
