+++ flatpak/platform-info.json
@@ -20,14 +20,14 @@
       "name": "platforminfo",
       "buildsystem": "simple",
       "build-commands": [
-        "install -D req2flatpak.py /app/bin/req2flatpak.py",
+        "install -D req2flatpak/req2flatpak.py /app/bin/req2flatpak.py",
         "echo '/usr/bin/env python3 /app/bin/req2flatpak.py --platform-info' > /app/bin/platform-info",
         "chmod +x /app/bin/platform-info"
       ],
       "sources": [
         {
           "type": "file",
-          "path": "req2flatpak.py"
+          "path": "req2flatpak/req2flatpak.py"
         }
       ]
     }
